Pulsed plane-wave fields (#14)
This PR adds an abstract interface for dealing with pulsed plane-wave
fields in the context of pulsed-perturbative QED. As the root type of
all background fields, to be used in the strong field part of `QED.jl`,
we introduce the abstract base type `AbstractBackgroundField`. However,
the interface is currently defined for pulsed plane-wave fields (see
below).

## Definition of the interface

The base type of the interface is `AbstractPulsedPlaneWaveFields`, where
the following interface functions are mandatory

```Julia
reference_momentum(::AbstractPulsedPlaneWaveFields) # return the reference momentum of the field
domain(::AbstractPulsedPlaneWaveFields) # return the interval, on which the BG field is non-zero
phase_duration(::AbstractPulsedPlaneWaveFields) # return the extent of the BG field in the phase variable phi
_pulse_envelope(::AbstractPulsedPlaneWaveFields, phi::Real) # return the value of the envelope function of the BG field in the phase variable phi, i.e. the phase profile 
```

The first three functions are just accessor functions for static
properties of the background field. The last function `_pulse_envelope`
defines the shape of the pulse, and it only needs to be implemented on
one single value of $\phi$. With that, the following functions were
implemented as well

```Julia
pulse_envelope(::AbstractPulsedPlaneWaveFields,phi::Real) # return the value of the envelope function at phi if phi is in the domain, and zero otherwise
_amplitude(::AbstractPulsedPlaneWaveFields, pol::AbstactDefinitePolarization,phi::Real) # return envelope times oscillator for given polarisation direction
amplitude(::AbstractPulsedPlaneWaveFields, pol::AbstactDefinitePolarization,phi::Real) # same as _amplitude, but return zero if phi is not in the domain
generic_spectrum(::AbstractPulsedPlaneWaveFields,pnum::Real) # return the Fourier transform of the amplitude at a given pnum
```

For each function, which is defined for `phi::Real` or `pnum::Real`, the
respective version for multiple values, i.e.
`phi::AbstractVector{<:Real}` and `pnum::AbstractVector{<:Real}` is also
implemented.

For the generic spectrum, the implementation is simply based on `QuadGK`
and therefore worthy of improvement.

---------

Co-authored-by: Uwe Hernandez Acosta <u.hernandez@hzdr.de>